+++ src/global.php
@@ -22,14 +22,3 @@ if (!function_exists('json5_decode')) {
         return \ColinODell\Json5\Json5Decoder::decode($json, $associative, $depth, $flags);
     }
 }
-
-// PHP 7.3 polyfills
-if (!\defined('JSON_THROW_ON_ERROR')) {
-    \define('JSON_THROW_ON_ERROR', 1 << 22);
-}
-
-if (!\class_exists('JsonException')) {
-    class JsonException extends Exception
-    {
-    }
-}